Address review comments: refactor structure, extract constants
- Split InspectionExpandedSection into ConversionPhaseLabel,
  InspectionTableRow, and slim orchestrator
- Consolidate types in utils/types.ts: VmOverrides, VmInspectionRef,
  DiskEncryptionParam, InspectionVmRowData
- Extract getInspectionStatusConfig to utils file
- Remove dead timestamp/showTimestamp code from InspectionStatusLabel
- Use getCategoryStatus from shared Concerns utility instead of
  inline getConcernCategoryStatus
- Extract resolveDiskEncryption from modal to utility function
- Use ConversionModel for apiVersion/kind in buildConversionCR
- Use CATEGORY_TYPES.READY and CONDITION_STATUS.TRUE constants
- Extract TOOLTIP_TRIGGER_MANUAL constant
- Add generic getLabels selector to common selectors
- Rename inspectionVmFields.tsx to .ts (no JSX)


Resolves: MTV-2482
