+++ src/components/InspectVirtualMachines/InspectVirtualMachinesModal.tsx
@@ -7,22 +7,17 @@ import type { V1beta1Plan, V1beta1Provider } from '@forklift-ui/types';
 import type { ModalComponent } from '@openshift-console/dynamic-plugin-sdk/lib/app/modal-support/ModalProvider';
 import { Alert, AlertVariant, ModalVariant } from '@patternfly/react-core';
 import { getNamespace, getUID, getVddkInitImage } from '@utils/crds/common/selectors';
-import {
-  CONVERSION_LABELS,
-  CONVERSION_TYPE,
-  DISK_ENCRYPTION_TYPE,
-} from '@utils/crds/conversion/constants';
+import { CONVERSION_LABELS, CONVERSION_TYPE } from '@utils/crds/conversion/constants';
 import { isEmpty } from '@utils/helpers';
 import { useInventoryVms } from '@utils/hooks/useInventoryVms';
 import { useVmInspectionStatus } from '@utils/hooks/useVmInspectionStatus';
 import { useWatchConversions } from '@utils/hooks/useWatchConversions';
 
-import type { VmInspectionRef } from './hooks/useCreateDeepInspections';
 import { useCreateDeepInspections } from './hooks/useCreateDeepInspections';
-import { createInspectionSecret } from './utils/createInspectionSecret';
 import { normalizeInventoryVms, normalizePlanVms } from './utils/normalizeVmsForInspection';
+import { resolveDiskEncryption } from './utils/resolveDiskEncryption';
+import type { VmInspectionRef, VmOverrides } from './utils/types';
 import InspectionVmTable from './InspectionVmTable';
-import type { VmOverrides } from './VmConfigForm';
 
 import './InspectVirtualMachinesModal.scss';
 
@@ -87,37 +82,15 @@ const InspectVirtualMachinesModal: ModalComponent<InspectVirtualMachinesModalPro
     setVmOverrides((prev) => ({ ...prev, [vmId]: overrides }));
   }, []);
 
-  const resolveDiskEncryption = useCallback(
-    async (vmId: string, vmName: string): Promise<VmInspectionRef['diskEncryption']> => {
-      const overrides = vmOverrides[vmId];
-      if (!overrides) return undefined;
-
-      if (overrides.nbdeClevis) {
-        return { type: DISK_ENCRYPTION_TYPE.CLEVIS };
-      }
-
-      const nonEmptyPhrases = (overrides.passphrases ?? []).filter((phrase) => !isEmpty(phrase));
-      if (!isEmpty(nonEmptyPhrases)) {
-        const secret = await createInspectionSecret(nonEmptyPhrases, vmName, namespace ?? '');
-        return {
-          secret: { name: secret.metadata?.name, namespace: secret.metadata?.namespace },
-          type: DISK_ENCRYPTION_TYPE.LUKS,
-        };
-      }
-
-      return undefined;
-    },
-    [vmOverrides, namespace],
-  );
-
   const handleConfirm = useCallback(async () => {
     const selectedSet = new Set(selectedIds);
     const selectedVms = vmRows.filter((vm) => selectedSet.has(vm.id) && !vm.isActive);
+    const ns = namespace ?? '';
 
     const vmsToInspect: VmInspectionRef[] = await Promise.all(
       selectedVms.map(async (vm) => {
         const diskEncryption = isProviderFlow
-          ? await resolveDiskEncryption(vm.id, vm.name)
+          ? await resolveDiskEncryption(vmOverrides[vm.id], vm.name, ns)
           : undefined;
         return { diskEncryption, id: vm.id, name: vm.name };
       }),
@@ -132,7 +105,7 @@ const InspectVirtualMachinesModal: ModalComponent<InspectVirtualMachinesModalPro
 
     setSelectedIds([]);
     setVmOverrides({});
-  }, [vmRows, selectedIds, createInspections, isProviderFlow, resolveDiskEncryption, t]);
+  }, [vmRows, selectedIds, vmOverrides, namespace, createInspections, isProviderFlow, t]);
 
   const isSubmitDisabled =
     selectedCount === 0 || !isVddkConfigured || !conversionsLoaded || Boolean(conversionsError);
