+++ src/utils/constants.ts
@@ -2,6 +2,8 @@ import { t } from '@utils/i18n';
 
 export const EMPTY_MSG = '-';
 
+export const TOOLTIP_TRIGGER_MANUAL = 'manual';
+
 export const Namespace = {
   AllProjects: '#ALL_NS#',
   Default: 'default',
