+++ src/utils/crds/common/selectors.ts
@@ -10,6 +10,8 @@ export const getCreatedAt = (resource: K8sResourceCommon) => resource?.metadata?
 
 export const getUID = (resource: K8sResourceCommon) => resource?.metadata?.uid;
 
+export const getLabels = (resource: K8sResourceCommon) => resource?.metadata?.labels;
+
 export const getOwnerReference = (resource: K8sResourceCommon) =>
   resource?.metadata?.ownerReferences?.[0];
